% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_cent.R
\name{buffer_cent}
\alias{buffer_cent}
\title{Generate buffers from centroid points}
\usage{
buffer_cent(x, z)
}
\arguments{
\item{x}{A SpatVector or sf polygon object.}

\item{z}{Radius of buffer in kilometers.}
}
\value{
Buffers of centroids.
}
\description{
Generate buffers from centroid points
}
\examples{
library(rnaturalearth)
library(SpatialR)

br = vect(rnaturalearth::ne_states("Brazil", returnclass = "sf"))
plot(br)

buffer_br = buffer_cent(br, 100000)
plot(buffer_br, add = T)

}
