% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointcent.R
\name{pointcent}
\alias{pointcent}
\title{Generate centroid points from polygons}
\usage{
pointcent(x)
}
\arguments{
\item{x}{A SpatVector or sf polygon object.}
}
\value{
The points of centroids.
}
\description{
Generate centroid points from polygons
}
\examples{
library(rnaturalearth)
library(SpatialR)

br = vect(rnaturalearth::ne_states("Brazil", returnclass = "sf"))
plot(br)

point_br = pointcent(br)
plot(point_br, add = T)

}
