% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_cent.R
\name{poly_cent}
\alias{poly_cent}
\title{Getting centroid from polygons (Centroids in lat long)}
\usage{
poly_cent(x)
}
\arguments{
\item{x}{A SpatVector or sf polygon object.}
}
\value{
The table of centroids with longitude and latitude.
}
\description{
Getting centroid from polygons (Centroids in lat long)
}
\examples{
library(rnaturalearth)
library(SpatialR)

br = vect(rnaturalearth::ne_states("Brazil", returnclass = "sf"))
polycent(br)

}
