% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polycent.R
\name{polycent}
\alias{polycent}
\title{Getting centroid from polygons (Centroids in lat long)}
\usage{
polycent(x)
}
\arguments{
\item{x}{A SpatVector or sf polygon object.}
}
\value{
The table of centroids with longitude and latitude.
}
\description{
Getting centroid from polygons (Centroids in lat long)
}
\examples{
library(terra)
library(SpatialR)
br = vect("C:/Users/Workshop/Documents/Research/Programming/SpatialR/Shapes/Brazil_States.shp")
polycent(br)

}
