% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialR-package.R
\docType{package}
\name{SpatialR-package}
\alias{SpatialR}
\alias{SpatialR-package}
\title{SpatialR: A quick module for demonstrating spatial analysis in R.}
\description{
This module is used to lightly scratch the surface regarding the potential for using R the 
    import and analyze spatial data. We will show you how to import shape, point, and raster files, how to 
    run some basic statistics on your collected data, and how to generate interactive and 3D maps.
}
\author{
\strong{Maintainer}: Angelina Locker \email{angelina.sweeney@utexas.edu} (YOUR-ORCID-ID)

}
\keyword{internal}
